Added IHattrickRepository to include a method to access data by Hattrick ID. Renamed Get by Id method and Repository to GetById(int id) and Get by predicate method to Query on IRepository. Adapted all references to these methods to reflect changes. Removed SeniorTeams property on League App class since the relation didn't existed. Updated all Hattrick entities and parsers to change uint properties to long to use a same type across the entire app. Added a try catch block on copy to clipboard feature on FormToken to remediate a problem with some systems.

+++ Hyperar.HattrickUltimate.BusinessObjects/Hattrick/CheckToken/Root.cs
@@ -39,7 +39,7 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.Hattrick.CheckToken
         /// <summary>
         /// Gets or sets the User ID.
         /// </summary>
-        public uint User { get; set; }
+        public long User { get; set; }
 
         #endregion Public Properties
     }
